* commit 'e1a57cbb1c2752feb9315f748836dc44e58d2dc6':
  ape: Use the proper variable type

Merged-by: Michael Niedermayer <michaelni@gmx.at>

+++ libavformat/apetag.c
@@ -38,7 +38,7 @@ static int ape_tag_read_field(AVFormatContext *s)
 {
     AVIOContext *pb = s->pb;
     uint8_t key[1024], *value;
-    uint32_t size, flags;
+    int64_t size, flags;
     int i, c;
 
     size = avio_rl32(pb);  /* field size */
